+++ shared/framework-kernel/src/declarations/framework/kernel.ts
@@ -15,4 +15,5 @@ export interface KernelOptions {
   variables: VariableOptions;
   staticRoutes: Array<RouteRecordRaw>;
   config: HttpConfig;
+  signOutExtension: () => void;
 }