+++ shared/framework-kernel/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @herodotus-cloud/framework-kernel
 
+## 1.7.10
+
+### Patch Changes
+
+- 调整代码错误
+- Updated dependencies
+  - @herodotus-cloud/core@1.7.10
+
+## 1.7.9
+
+### Patch Changes
+
+- 迁移最后的管理代码
+- Updated dependencies
+  - @herodotus-cloud/core@1.7.9
+
 ## 1.7.8
 
 ### Patch Changes